+++ build.gradle
@@ -45,7 +45,7 @@ dependencies {
 	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5' // JSON 파싱용
 
 	// aws s3
-	implementation 'io.awspring.cloud:spring-cloud-aws-starter-s3:3.1.0'
+	implementation 'io.awspring.cloud:spring-cloud-aws-starter-s3:3.4.0'
 
 	// db
 	runtimeOnly 'com.mysql:mysql-connector-j'
