+++ src/main/java/com/umc/nuvibe/domain/image/service/ImageService.java
@@ -9,25 +9,8 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-@Service
-@AllArgsConstructor
-public class ImageService {
 
-    private final S3Service s3Service;
-    private final ImageRepository imageRepository;
-
-    public ImageRes uploadAndSave(MultipartFile file, ImageTag tag) {
-
-        String imageURL= s3Service.upload(file);
-
-        Image newImage = Image.builder()
-                .imageUrl(imageURL)
-                .imageTag(tag)
-                .build();
-
-        imageRepository.save(newImage);
-
-        return new ImageRes(imageURL, tag);
-    }
+public interface ImageService {
+    ImageRes uploadAndSave(MultipartFile file, ImageTag tag);
 
 }
