+++ src/main/java/com/umc/nuvibe/global/apiPayLoad/error/ImageErrorCode.java
@@ -13,6 +13,7 @@ public enum ImageErrorCode implements ErrorCode {
     INVALID_FILE_EXTENSION(HttpStatus.BAD_REQUEST, "COMMON002","옳바르지 않은 확장자입니다."),
     EXTENSION_IS_EMPTY(HttpStatus.BAD_REQUEST, "IMAGE003","확장자가 존재하지 않습니다."),
     IMAGE_UPLOAD_FAIL(HttpStatus.BAD_REQUEST, "IMAGE004","이미지 업로드에 실패했습니다"),
+    IMAGETAG_IS_NULL(HttpStatus.BAD_REQUEST,"IMAGE005","이미지 태그가 입력되지 않았습니다"),
 
     ;
